fix(post): change status type from uint8 to int64 in models and update related structures for consistency

+++ internal/module/post/models/constants.go
@@ -1,7 +1,7 @@
 package models
 
 const (
-	STATUS_DRAFT uint8 = iota
+	STATUS_DRAFT int64 = iota
 	STATUS_PENDING
 	STATUS_APPROVED
 	STATUS_REJECTED
